feat: 🎸 추가 요일 관련 생성자 추가, 영업시간/휴업일 repository test code 추가
testEntity 의 사용되지 않던 getHoliday 함수 삭제
