+++ src/test/java/store/cookshoong/www/cookshoongbackend/store/repository/businesshour/BusinessHourRepositoryTest.java
@@ -1,7 +1,7 @@
-package store.cookshoong.www.cookshoongbackend.store.repository;
+package store.cookshoong.www.cookshoongbackend.store.repository.businesshour;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -11,20 +11,20 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
-import store.cookshoong.www.cookshoongbackend.store.entity.Holiday;
+import store.cookshoong.www.cookshoongbackend.store.entity.BusinessHour;
+import store.cookshoong.www.cookshoongbackend.store.entity.DayType;
 import store.cookshoong.www.cookshoongbackend.store.entity.Store;
-import store.cookshoong.www.cookshoongbackend.store.repository.businesshour.HolidayRepository;
 import store.cookshoong.www.cookshoongbackend.util.TestPersistEntity;
 
 @DataJpaTest
 @Import(TestPersistEntity.class)
-class HolidayRepositoryTest {
+class BusinessHourRepositoryTest {
 
     @MockBean
     JPAQueryFactory jpaQueryFactory;
 
     @Autowired
-    HolidayRepository holidayRepository;
+    BusinessHourRepository businessHourRepository;
 
     @Autowired
     TestPersistEntity tpe;
@@ -33,17 +33,16 @@ class HolidayRepositoryTest {
     TestEntityManager em;
 
     @Test
-    @DisplayName("휴업일 조회 - 성공")
-    void find_holiday() {
+    @DisplayName("영업시간 조회 - 성공")
+    void findByStore_Id() {
         Store store = tpe.getOpenStore();
-        Holiday actual = new Holiday(store, LocalDate.of(2020, 2, 20), LocalDate.of(2020, 2, 22));
-        holidayRepository.save(actual);
-
+        DayType dayType = new DayType("MON", "월요일");
+        BusinessHour actual = new BusinessHour(store, dayType, LocalTime.of(9,00), LocalTime.of(21,00));
+        em.persist(dayType);
+        businessHourRepository.save(actual);
         em.clear();
 
-        List<Holiday> expect = holidayRepository.findByStore_Id(store.getId());
+        List<BusinessHour> expect = businessHourRepository.findByStore_Id(store.getId());
         Assertions.assertThat(expect).hasSize(1);
     }
-
-
 }
